WIP: Add SpiralTreeMapLayoutGenerator
Which is better able to manage minimum cell sizes by doing layout in a spiral.
Does not fail gracefully right now.
TODO: need to add some sort of retracing so that it doesn't blindly pick the next one that fits.
Ends up with weird gaps that can't be fit.
Also something is wrong with the availableBounds logic